feat: add a new filter config class for the zeebe gateway
Adds a new FilterCfg class for the Zeebe Gateway. The
class will be used to provide the configuration for
loading a single custom implementation of a Jakarta
Servlet Filter. These filters will be used as the
custom middleware for the Zeebe REST API.

The List of FilterCfg instances is contained within
the GatewayCfg class, which will provide these config
properties to the Zeebe REST API for instantiation.

The FilterCfg class is a copy of the InterceptorCfg
class since it provides the same functionality but
for Filters instead of (gRPC) Interceptors. Re-using
the same code was considered to avoid duplication,
but this might hurt us in the future if we decide to
expand the Filter configuration properties and not
the Interceptor properties.