refactor: use a list instead of a map with custom interceptors
The InterceptorRepository class loads and keeps custom
interceptors. The interceptors should be kept in the
order defined in the Zeebe configuration. To ensure the
correct order, a SortedMap was used. However, this approach
made the code less readable and more complex.

This commit refactors the InterceptorRepository class to
use an ArrayList to ensure the correct ordering of custom
interceptors.