feat: process message correlate command for msg start events
Adds the MessageCorrelationProcessor and processes the MessageCorrelation CORRELATE command. It works similar to the
Message PUBLISH command. I have chosen to keep the message lifecycle mostly the same for clarity. Especially since
consumers of exported records might expect some events to happen.

The processor will first mark the message as published. Next it will find any start event subscriptions that match the
message. If it does it starts the PI. It'll take the first one it started and use this PI key for the response back to
the client.
If no subscriptions are found it'll respond with a NOT_CORRELATED message.
Finally the message is expired immediately, as no TTL is set.